reorganize modules (#512)
* ingest permissionList module

* merge all permissions modules

* merge review modules

* more cleanup

* merge developer modules

* remove similar default parameters object

* remove similar mapper

* delete all mappers

* use base_url consistently

* move processPages util

* fix
