+++ lib/categories.js
@@ -2,8 +2,9 @@
 
 const request = require('./utils/request');
 const cheerio = require('cheerio');
+const { BASE_URL } = require('./constants');
 
-const PLAYSTORE_URL = 'https://play.google.com/store/apps';
+const PLAYSTORE_URL = `${BASE_URL}/store/apps`;
 const CATEGORY_URL_PREFIX = '/store/apps/category/';
 
 function categories (opts) {
