+++ lib/constants.js
@@ -1,5 +1,7 @@
 'use strict';
 
+module.exports.BASE_URL = 'https://play.google.com';
+
 module.exports.clusters = {
   new: 'new',
   top: 'top'
