+++ lib/developer.js
@@ -3,9 +3,10 @@
 const debug = require('debug')('google-play-scraper:developer');
 const qs = require('querystring');
 const scriptData = require('./utils/scriptData');
-const parseDeveloperApps = require('./utils/parseDeveloperApps');
-const { BASE_URL } = require('./utils/configurations');
+const { BASE_URL } = require('./constants');
 const request = require('./utils/request');
+const R = require('ramda');
+const processPages = require('./utils/processPages');
 
 function buildUrl (opts) {
   const { lang, devId, country } = opts;
@@ -51,6 +52,42 @@ function developer (opts) {
       .then(resolve)
       .catch(reject);
   });
+}
+
+const INITIAL_MAPPINGS = {
+  cluster: ['ds:3', 0, 1, 0, 0, 3, 4, 2],
+  apps: ['ds:3', 0, 1, 0, 0, 0],
+  token: ['ds:3', 0, 1, 0, 0, 7, 1]
+};
+
+function parseDeveloperApps (developerObject, opts) {
+  const clusterUrl = opts.hasClusterUrl
+    ? getParsedCluster(developerObject)
+    : undefined;
+
+  if (!clusterUrl) {
+    return processPages(developerObject, opts, [], INITIAL_MAPPINGS);
+  }
+
+  const clusterUrlToProcess = `${BASE_URL}${clusterUrl}&hl=${opts.lang}&gl=${opts.country}`;
+
+  debug('Cluster Request URL: %s', clusterUrlToProcess);
+
+  const options = Object.assign({
+    url: clusterUrlToProcess,
+    method: 'GET',
+    followRedirect: true
+  }, opts.requestOptions);
+
+  return request(options, opts.throttle)
+    .then(scriptData.parse)
+    .then(clusterObject => processPages(clusterObject, opts, [], INITIAL_MAPPINGS))
+    .catch(console.error);
+}
+
+function getParsedCluster (developerObject) {
+  const clusterUrl = R.path(INITIAL_MAPPINGS.cluster, developerObject);
+  return clusterUrl;
 }
 
 module.exports = developer;
