+++ lib/list.js
@@ -6,8 +6,8 @@ const qs = require('querystring');
 const R = require('ramda');
 const c = require('./constants');
 const scriptData = require('./utils/scriptData');
-const { processAndRecur } = require('./requesters/appsMappedRequests');
-const { BASE_URL } = require('./utils/configurations');
+const processPages = require('./utils/processPages');
+const { BASE_URL } = require('./constants');
 
 function list (opts) {
   return new Promise(function (resolve, reject) {
@@ -154,7 +154,7 @@ function parseCategoryApps (categoryObject, opts) {
 
   return request(options, opts.throttle)
     .then(scriptData.parse)
-    .then(clusterObject => processAndRecur(clusterObject, opts, [], APP_MAPPINGS))
+    .then(clusterObject => processPages(clusterObject, opts, [], APP_MAPPINGS))
     .catch(console.error);
 }
