--- lib/mappers/apps.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const REQUEST_MAPPINGS = {
-  apps: [0, 0, 0],
-  token: [0, 0, 7, 1]
-};
-
-module.exports = {
-  REQUEST_MAPPINGS
-};
