--- lib/mappers/permissions.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const MAPPINGS = {
-  permissions: [2],
-  type: 0
-};
-
-module.exports = {
-  MAPPINGS
-};
