--- lib/mappers/reviews.js
@@ -1,16 +1,0 @@
-'use strict';
-
-const INITIAL_MAPPINGS = {
-  reviews: ['ds:15', 0],
-  token: ['ds:15', 1, 1]
-};
-
-const REQUEST_MAPPINGS = {
-  reviews: [0],
-  token: [1, 1]
-};
-
-module.exports = {
-  INITIAL_MAPPINGS,
-  REQUEST_MAPPINGS
-};
