--- lib/mappers/similar.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const INITIAL_MAPPINGS = {
-  cluster: ['ds:7', 1, 1, 0, 0, 3, 4, 2],
-  apps: ['ds:3', 0, 1, 0, 0, 0],
-  token: ['ds:3', 0, 1, 0, 0, 7, 1]
-};
-
-module.exports = {
-  INITIAL_MAPPINGS
-};
