+++ lib/search.js
@@ -4,6 +4,7 @@ const request = require('./utils/request');
 const scriptData = require('./utils/scriptData');
 const appList = require('./utils/appList');
 const R = require('ramda');
+const { BASE_URL } = require('./constants');
 
 const body = 'f.req=%5B%5B%5B%22qnKhOb%22%2C%22%5B%5Bnull%2C%5B%5B10%2C%5B10%2C50%5D%5D%2Ctrue%2Cnull%2C%5B96%2C27%2C4%2C8%2C57%2C30%2C110%2C79%2C11%2C16%2C49%2C1%2C3%2C9%2C12%2C104%2C55%2C56%2C51%2C10%2C34%2C77%5D%5D%2Cnull%2C%5C%22%token%%5C%22%5D%5D%22%2Cnull%2C%22generic%22%5D%5D%5D';
 
@@ -11,6 +12,7 @@ const body = 'f.req=%5B%5B%5B%22qnKhOb%22%2C%22%5B%5Bnull%2C%5B%5B10%2C%5B10%2C5
  * Extract navigation tokens for next pages, parse results and call
  * `checkFinished` to repeat the process with next page if necessary.
  */
+// FIXME this looks awfully similar to the other processAndRecur
 function processAndRecur (html, opts, savedApps, mappings) {
   if (R.is(String, html)) {
     html = scriptData.parse(html);
@@ -33,7 +35,7 @@ function checkFinished (opts, savedApps, nextToken) {
   }
 
   const requestOptions = Object.assign({
-    url: `https://play.google.com/_/PlayStoreUi/data/batchexecute?rpcids=qnKhOb&bl=boq_playuiserver_20190424.04_p0&hl=${opts.lang}&gl=${opts.country}&authuser=0&soc-app=121&soc-platform=1&soc-device=1`,
+    url: `${BASE_URL}/_/PlayStoreUi/data/batchexecute?rpcids=qnKhOb&bl=boq_playuiserver_20190424.04_p0&hl=${opts.lang}&gl=${opts.country}&authuser=0&soc-app=121&soc-platform=1&soc-device=1`,
     method: 'POST',
     body: body.replace('%token%', nextToken),
     followRedirect: true,
@@ -60,7 +62,7 @@ function initialRequest (opts) {
   function skipClusterPage (html) {
     const match = html.match(/href="\/store\/apps\/collection\/search_collection_more_results_cluster?(.*?)"/);
     if (match) {
-      const innerUrl = 'https://play.google.com/' + match[0].split(/"/)[1];
+      const innerUrl = BASE_URL + match[0].split(/"/)[1];
       return request(Object.assign({
         url: innerUrl
       }, opts.requestOptions), opts.throttle);
@@ -68,7 +70,7 @@ function initialRequest (opts) {
     return html;
   }
 
-  const url = `https://play.google.com/store/search?c=apps&q=${opts.term}&hl=${opts.lang}&gl=${opts.country}&price=${opts.price}`;
+  const url = `${BASE_URL}/store/search?c=apps&q=${opts.term}&hl=${opts.lang}&gl=${opts.country}&price=${opts.price}`;
   return request(Object.assign({ url }, opts.requestOptions), opts.throttle)
     .then(skipClusterPage)
     .then((html) => processAndRecur(html, opts, [], INITIAL_MAPPINGS));
