+++ lib/similar.js
@@ -5,9 +5,8 @@ const debug = require('debug')('google-play-scraper:similar');
 const request = require('./utils/request');
 const queryString = require('querystring');
 const scriptData = require('./utils/scriptData');
-const { BASE_URL, DEFAULT_PARAMETERS } = require('./utils/configurations');
-const { processAndRecur } = require('./requesters/appsMappedRequests');
-const { INITIAL_MAPPINGS } = require('./mappers/similar');
+const { BASE_URL } = require('./constants');
+const processPages = require('./utils/processPages');
 
 function similar (opts) {
   return new Promise(function (resolve, reject) {
@@ -16,9 +15,9 @@ function similar (opts) {
     const mergedOpts = Object.assign({},
       {
         appId: encodeURIComponent(opts.appId),
-        lang: opts.lang || DEFAULT_PARAMETERS.similar.lang,
-        country: opts.country || DEFAULT_PARAMETERS.similar.country,
-        fullDetail: opts.fullDetail || DEFAULT_PARAMETERS.similar.fullDetail
+        lang: opts.lang || 'en',
+        country: opts.country || 'us',
+        fullDetail: opts.fullDetail
       });
 
     const qs = queryString.stringify({
@@ -49,6 +48,12 @@ function validateSimilarParameters (opts) {
   }
 }
 
+const INITIAL_MAPPINGS = {
+  cluster: ['ds:7', 1, 1, 0, 0, 3, 4, 2],
+  apps: ['ds:3', 0, 1, 0, 0, 0],
+  token: ['ds:3', 0, 1, 0, 0, 7, 1]
+};
+
 function parseSimilarApps (similarObject, opts) {
   const clusterUrl = getParsedCluster(similarObject);
 
@@ -66,7 +71,7 @@ function parseSimilarApps (similarObject, opts) {
 
   return request(options, opts.throttle)
     .then(scriptData.parse)
-    .then((htmlParsed) => processAndRecur(htmlParsed, opts, [], INITIAL_MAPPINGS));
+    .then((htmlParsed) => processPages(htmlParsed, opts, [], INITIAL_MAPPINGS));
 }
 
 function getParsedCluster (similarObject) {
