+++ lib/utils/appList.js
@@ -3,13 +3,14 @@
 const url = require('url');
 const R = require('ramda');
 const scriptData = require('./scriptData');
+const { BASE_URL } = require('../constants');
 
 const MAPPINGS = {
   title: [2],
   appId: [12, 0],
   url: {
     path: [9, 4, 2],
-    fun: (path) => new url.URL(path, 'https://play.google.com').toString()
+    fun: (path) => new url.URL(path, BASE_URL).toString()
   },
   icon: [1, 1, 0, 3, 2],
   developer: [4, 0, 0, 0],
