--- lib/utils/configurations.js
@@ -1,16 +1,0 @@
-'use strict';
-
-const BASE_URL = 'https://play.google.com';
-
-const DEFAULT_PARAMETERS = {
-  similar: {
-    lang: 'en',
-    country: 'us',
-    fullDetail: false
-  }
-};
-
-module.exports = {
-  BASE_URL,
-  DEFAULT_PARAMETERS
-};
