+++ lib/utils/parseList.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const { BASE_URL } = require('../constants');
+
 /*
  * Return the proper parseList function according to the options.
  */
@@ -52,7 +54,7 @@ function parseApp (app) {
   }
 
   return {
-    url: 'https://play.google.com' + app.find('a').attr('href'),
+    url: BASE_URL + app.find('a').attr('href'),
     appId: app.attr('data-docid'),
     title: app.find('a.title').attr('title'),
     summary: app.find('div.description').text().trim(),
